Make it easy to select a Markdown executible.
For example, if you want to use a local version of Markdown to work
around bugs in your system's version.

The latest version of Gruber's Markdown seems to be:
  http://daringfireball.net/projects/downloads/Markdown_1.0.2b8.tbz
Release note:
  http://six.pairlist.net/pipermail/markdown-discuss/2007-May/000615.html
Debian experimental is using this 1.0.2b8 version:
  http://packages.debian.org/search?keywords=markdown

There are many other implementations as well, and most seem to support
the
  /path/to/markdown SOURCE
syntax, printing the marked up XHTML to stdout.  This patch makes it
easy to drop in any executable you wish.